packet dump EDA - error resolved
There was an error during using the actual generator variable which has the output of TimeSeriesGenerator while implementing the FG1&FG3 functions. After bunch of debuggings, I found out that instead of the feature generator functions, there was a miss in the TimeSeriesGenerator class, that there wasn't any functions processing the iterables. (such as for X in generator: X / the original class only had functions processing types such as for i in range(len(generator)): generator[i]. These two seems to work same, but it has a difference.) So I add the __iter__ magic method for this, and finally the error has been resolved!